+++ onmt/decoders/transformer.py
@@ -10,7 +10,11 @@ from onmt.modules import MultiHeadedAttention, AverageAttention
 from onmt.modules.position_ffn import PositionwiseFeedForward
 from onmt.modules.position_ffn import ActivationFunction
 from onmt.utils.misc import sequence_mask
-from onmt.modules.rmsnorm import RMSNorm
+
+try:
+    from apex.normalization import FusedRMSNorm as RMSNorm
+except ImportError:
+    from onmt.modules.rmsnorm import RMSNorm
 
 
 class TransformerDecoderLayerBase(nn.Module):
@@ -37,6 +41,7 @@ class TransformerDecoderLayerBase(nn.Module):
         norm_eps=1e-6,
         use_ckpting=[],
         parallel_gpu=1,
+        sliding_window=0,
     ):
         """
         Args:
@@ -115,8 +120,10 @@ class TransformerDecoderLayerBase(nn.Module):
             raise ValueError(f"{layer_norm} layer norm type is not supported")
 
         self.dropout = nn.Dropout(dropout)
+        self.dropout_p = dropout
         self.full_context_alignment = full_context_alignment
         self.alignment_heads = alignment_heads
+        self.sliding_window = sliding_window
 
     def forward(self, *args, **kwargs):
         """Extend `_forward` for (possibly) multiple decoder pass:
@@ -170,12 +177,12 @@ class TransformerDecoderLayerBase(nn.Module):
                 device=tgt_pad_mask.device,
                 dtype=torch.uint8,
             )
-            future_mask = future_mask.triu_(1).view(1, tgt_len, tgt_len)
-            # BoolTensor was introduced in pytorch 1.2
-            try:
-                future_mask = future_mask.bool()
-            except AttributeError:
-                pass
+            future_mask = future_mask.tril_(0)
+            if self.sliding_window > 0:
+                future_mask = future_mask.triu_(-self.sliding_window)
+            future_mask = future_mask.bool()
+            future_mask = ~future_mask.view(1, tgt_len, tgt_len)
+
             dec_mask = torch.gt(tgt_pad_mask + future_mask, 0)
         else:  # only mask padding, result mask in (B, 1, T)
             dec_mask = tgt_pad_mask
@@ -188,6 +195,7 @@ class TransformerDecoderLayerBase(nn.Module):
                 norm_layer_in,
                 norm_layer_in,
                 mask=dec_mask,
+                sliding_window=self.sliding_window,
                 step=step,
                 return_attn=return_attn,
             )
@@ -229,6 +237,7 @@ class TransformerDecoderLayer(TransformerDecoderLayerBase):
         norm_eps=1e-6,
         use_ckpting=[],
         parallel_gpu=1,
+        sliding_window=0,
     ):
         """
         Args:
@@ -256,6 +265,7 @@ class TransformerDecoderLayer(TransformerDecoderLayerBase):
             norm_eps=norm_eps,
             use_ckpting=use_ckpting,
             parallel_gpu=parallel_gpu,
+            sliding_window=sliding_window,
         )
         self.context_attn = MultiHeadedAttention(
             heads,
@@ -322,8 +332,11 @@ class TransformerDecoderLayer(TransformerDecoderLayerBase):
 
         norm_layer_in = self.layer_norm_1(layer_in)
 
-        self_attn, _ = self._forward_self_attn(norm_layer_in, dec_mask, step)
-
+        self_attn, _ = self._forward_self_attn(
+            norm_layer_in, dec_mask, step, return_attn=return_attn
+        )
+        if self.dropout_p > 0:
+            self_attn = self.dropout(self_attn)
         if self.parallel_residual:
             ctx_attn, attns = self.context_attn(
                 enc_out,
@@ -337,16 +350,18 @@ class TransformerDecoderLayer(TransformerDecoderLayerBase):
                 self.feed_forward(norm_layer_in)
                 - norm_layer_in
                 + layer_in
-                + self.dropout(self_attn)
+                + self_attn
                 + ctx_attn
             )
         else:
-            query = self.dropout(self_attn) + layer_in
+            query = self_attn + layer_in
             norm_query = self.layer_norm_2(query)
             ctx_attn, attns = self.context_attn(
                 enc_out, enc_out, norm_query, mask=src_pad_mask, return_attn=return_attn
             )
-            layer_out = self.feed_forward(self.dropout(ctx_attn) + query)
+            if self.dropout_p > 0:
+                ctx_attn = self.dropout(ctx_attn)
+            layer_out = self.feed_forward(ctx_attn + query)
 
         return layer_out, attns
 
@@ -408,6 +423,7 @@ class TransformerDecoderBase(DecoderBase):
             parallel_gpu=opt.world_size
             if opt.parallel_mode == "tensor_parallel"
             else 1,
+            sliding_window=opt.sliding_window,
         )
 
     def init_state(self, src, enc_out, enc_final_hs):
@@ -501,6 +517,7 @@ class TransformerDecoder(TransformerDecoderBase):
         norm_eps=1e-6,
         use_ckpting=[],
         parallel_gpu=1,
+        sliding_window=0,
     ):
         super(TransformerDecoder, self).__init__(
             d_model, copy_attn, embeddings, alignment_layer, layer_norm, norm_eps
@@ -530,6 +547,7 @@ class TransformerDecoder(TransformerDecoderBase):
                     norm_eps=norm_eps,
                     use_ckpting=use_ckpting,
                     parallel_gpu=parallel_gpu,
+                    sliding_window=sliding_window,
                 )
                 for i in range(num_layers)
             ]
@@ -575,7 +593,9 @@ class TransformerDecoder(TransformerDecoderBase):
         tgt_pad_mask = tgt[:, :, 0].eq(pad_idx).unsqueeze(1)  # [B, 1, T_tgt]
 
         with_align = kwargs.pop("with_align", False)
-        return_attn = with_align or self._copy
+        return_attn = kwargs.pop("return_attn", False)
+        return_attn = with_align or self._copy or return_attn
+
         attn_aligns = []
 
         for layer in self.transformer_layers:
@@ -631,6 +651,8 @@ class TransformerDecoder(TransformerDecoderBase):
                         "values": torch.tensor([], device=enc_out.device),
                     },
                 )
+                if hasattr(layer.self_attn, "rope"):
+                    layer.self_attn.rope = layer.self_attn.rope.to(enc_out.device)
 
 
 class TransformerLMDecoderLayer(TransformerDecoderLayerBase):
@@ -676,7 +698,8 @@ class TransformerLMDecoderLayer(TransformerDecoderLayerBase):
         attn_output, attns = self._forward_self_attn(
             norm_layer_in, dec_mask, step, return_attn=return_attn
         )
-
+        if self.dropout_p > 0:
+            attn_output = self.dropout(attn_output)
         if self.parallel_residual:
             # feed_forward applies residual, so we remove and apply residual with un-normed
             if not self.shared_layer_norm:
@@ -684,11 +707,9 @@ class TransformerLMDecoderLayer(TransformerDecoderLayerBase):
                 ff_in = norm_res_layer_in
             else:
                 ff_in = norm_layer_in
-            layer_out = (
-                self.feed_forward(ff_in) - ff_in + layer_in + self.dropout(attn_output)
-            )
+            layer_out = self.feed_forward(ff_in) - ff_in + layer_in + attn_output
         else:
-            layer_out = self.dropout(attn_output) + layer_in
+            layer_out = attn_output + layer_in
             layer_out = self.feed_forward(layer_out)
 
         return layer_out, attns
@@ -742,6 +763,7 @@ class TransformerLMDecoder(TransformerDecoderBase):
         norm_eps=1e-6,
         use_ckpting=[],
         parallel_gpu=1,
+        sliding_window=0,
     ):
         super(TransformerLMDecoder, self).__init__(
             d_model, copy_attn, embeddings, alignment_layer, layer_norm, norm_eps
@@ -770,6 +792,7 @@ class TransformerLMDecoder(TransformerDecoderBase):
                     norm_eps=norm_eps,
                     use_ckpting=use_ckpting,
                     parallel_gpu=parallel_gpu,
+                    sliding_window=sliding_window,
                 )
                 for i in range(num_layers)
             ]
@@ -800,7 +823,8 @@ class TransformerLMDecoder(TransformerDecoderBase):
         tgt_pad_mask = tgt[:, :, 0].eq(pad_idx).unsqueeze(1)  # [B, 1, T_tgt]
 
         with_align = kwargs.pop("with_align", False)
-        return_attn = with_align or self._copy
+        return_attn = kwargs.pop("return_attn", False)
+        return_attn = with_align or self._copy or return_attn
         assert not with_align, "TransformerLMDecoder does not support align"
 
         for layer in self.transformer_layers:
@@ -833,3 +857,5 @@ class TransformerLMDecoder(TransformerDecoderBase):
                         "values": torch.tensor([], device=tgt.device),
                     },
                 )
+                if hasattr(layer.self_attn, "rope"):
+                    layer.self_attn.rope = layer.self_attn.rope.to(tgt.device)