+++ onmt/model_builder.py
@@ -120,7 +120,9 @@ def load_test_model(opt, device_id=0, model_path=None):
 
     # Avoid functionality on inference
     model_opt.update_vocab = False
-
+    model_opt.attention_dropout = (
+        0.0  # required to force no dropout at inference with flash
+    )
     model = build_base_model(model_opt, vocabs)
 
     precision = torch.float32
@@ -157,7 +159,9 @@ def load_test_model(opt, device_id=0, model_path=None):
     del checkpoint
 
     model.eval()
-    model.generator.eval()
+    for name, module in model.named_modules():
+        if hasattr(module, "dropout_p"):
+            module.dropout_p = 0.0
     return vocabs, model, model_opt